fetch/metadata/generated/worker-dedicated-importscripts.https.sub.tentative.html

<!DOCTYPE html>
<!--
This test was procedurally generated. Please do not modify it directly.
Sources:
- fetch/metadata/tools/fetch-metadata.conf.yml
- fetch/metadata/tools/templates/worker-dedicated-importscripts.sub.html
-->
<html lang="en">
  <meta charset="utf-8">
  <title>HTTP headers on request for dedicated worker via the "importScripts" API</title>
  <script src="/resources/testharness.js"></script>
  <script src="/resources/testharnessreport.js"></script>
  <script src="/fetch/metadata/resources/helper.sub.js"></script>
  <script type="module">
  'use strict';
  function induceRequest(url, options) {
    const src = `
      importScripts('${url}');
      postMessage('done');
    `;
    const workerUrl = URL.createObjectURL(
      new Blob([src], { type: 'application/javascript' })
    );
    return new Promise((resolve, reject) => {
      const worker = new Worker(workerUrl, options);
      worker.onmessage = resolve;
      worker.onerror = reject;
    });
  }

  promise_test(() => {
    const key = '{{uuid()}}';
    const url = makeRequestURL(
      key, ['httpsOrigin'], { mime: 'application/javascript' }
    );

    return induceRequest(url)
      .then(() => retrieve(key))
      .then((headers) => {
          assert_own_property(headers, 'sec-fetch-frame-ancestors');
          assert_array_equals(headers['sec-fetch-frame-ancestors'], ['same-origin']);
        });
  }, 'sec-fetch-frame-ancestors - Same origin');

  promise_test(() => {
    const key = '{{uuid()}}';
    const url = makeRequestURL(
      key, ['httpsCrossSite'], { mime: 'application/javascript' }
    );

    return induceRequest(url)
      .then(() => retrieve(key))
      .then((headers) => {
          assert_own_property(headers, 'sec-fetch-frame-ancestors');
          assert_array_equals(headers['sec-fetch-frame-ancestors'], ['cross-site']);
        });
  }, 'sec-fetch-frame-ancestors - Cross-site');

  promise_test(() => {
    const key = '{{uuid()}}';
    const url = makeRequestURL(
      key, ['httpsSameSite'], { mime: 'application/javascript' }
    );

    return induceRequest(url)
      .then(() => retrieve(key))
      .then((headers) => {
          assert_own_property(headers, 'sec-fetch-frame-ancestors');
          assert_array_equals(headers['sec-fetch-frame-ancestors'], ['same-site']);
        });
  }, 'sec-fetch-frame-ancestors - Same site');
  </script>
</html>
